backend/pyproject.toml
[tool.poetry]
name = "vait-govhack-melb-backend"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8"
ariadne = "^0.13.0"
boto3 = "^1.18.17"
Flask = "^2.0.1"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.19.0"
SQLAlchemy = "^1.4.23"
ipykernel = "^6.2.0"
geopandas = "^0.9.0"
psycopg2-binary = "^2.9.1"
PyMySQL = "^1.0.2"
fsspec = "^2021.7.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
